# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "安信合同管理系统 API",
    "description": "基于 Go + Gin + MySQL 的合同管理系统 RESTful API 文档",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "host": "localhost:8000",
  "basePath": "/",
  "produces": ["application/json"],
  "consumes": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Token. Example: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/api/contracts": {
      "get": {
        "summary": "获取合同列表",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "query", "name": "skip", "type": "integer"},
          {"in": "query", "name": "limit", "type": "integer"},
          {"in": "query", "name": "customer_id", "type": "integer"},
          {"in": "query", "name": "contract_type_id", "type": "integer"},
          {"in": "query", "name": "status", "type": "string"}
        ],
        "responses": {
          "200": {"description": "成功"}
        }
      },
      "post": {
        "summary": "创建合同",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["title", "customer_id", "contract_type_id"],
              "properties": {
                "title": {"type": "string"},
                "customer_id": {"type": "integer"},
                "contract_type_id": {"type": "integer"},
                "amount": {"type": "number"},
                "currency": {"type": "string", "default": "CNY"},
                "sign_date": {"type": "string", "format": "date"},
                "start_date": {"type": "string", "format": "date"},
                "end_date": {"type": "string", "format": "date"},
                "payment_terms": {"type": "string"},
                "content": {"type": "string"},
                "notes": {"type": "string"}
              }
            }
          }
        ],
        "responses": {
          "201": {"description": "创建成功"}
        }
      }
    },
    "/api/contracts/{contract_id}": {
      "get": {
        "summary": "获取合同详情",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "path", "name": "contract_id", "type": "integer", "required": true}
        ],
        "responses": {
          "200": {"description": "成功"}
        }
      },
      "put": {
        "summary": "更新合同",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "path", "name": "contract_id", "type": "integer", "required": true}
        ],
        "responses": {
          "200": {"description": "成功"}
        }
      },
      "delete": {
        "summary": "删除合同",
        "tags": ["合同管理"],
        "security": [{"Bearer": []}],
        "parameters": [
          {"in": "path", "name": "contract_id", "type": "integer", "required": true}
        ],
        "responses": {
          "204": {"description": "删除成功"}
        }
      }
    },
...
  },
...
}